Add chan buffer size support
